refactor: next.config.js 리팩터링
