+++ package.json
@@ -38,6 +38,7 @@
     "axios": "^1.2.0",
     "lottie-web": "^5.10.2",
     "next": "13.0.5",
+    "next-compose-plugins": "^2.2.1",
     "next-pwa": "^5.6.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
